+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidOptionsInitializer.java
@@ -12,9 +12,9 @@ import io.sentry.SendCachedEnvelopeFireAndForgetIntegration;
 import io.sentry.SendFireAndForgetEnvelopeSender;
 import io.sentry.SendFireAndForgetOutboxSender;
 import io.sentry.SentryLevel;
-import io.sentry.SentryOptions;
 import io.sentry.android.fragment.FragmentLifecycleIntegration;
 import io.sentry.android.timber.SentryTimberIntegration;
+import io.sentry.util.FileUtils;
 import io.sentry.util.Objects;
 import java.io.BufferedInputStream;
 import java.io.File;
@@ -27,7 +27,7 @@ import org.jetbrains.annotations.Nullable;
 
 /**
  * Android Options initializer, it reads configurations from AndroidManifest and sets to the
- * SentryOptions. It also adds default values for some fields.
+ * SentryAndroidOptions. It also adds default values for some fields.
  */
 @SuppressWarnings("Convert2MethodRef") // older AGP versions do not support method references
 final class AndroidOptionsInitializer {
@@ -38,7 +38,7 @@ final class AndroidOptionsInitializer {
   /**
    * Init method of the Android Options initializer
    *
-   * @param options the SentryOptions
+   * @param options the SentryAndroidOptions
    * @param context the Application context
    */
   static void init(final @NotNull SentryAndroidOptions options, final @NotNull Context context) {
@@ -51,7 +51,7 @@ final class AndroidOptionsInitializer {
   /**
    * Init method of the Android Options initializer
    *
-   * @param options the SentryOptions
+   * @param options the SentryAndroidOptions
    * @param context the Application context
    * @param logger the ILogger interface
    * @param isFragmentAvailable whether the Fragment integration is available on the classpath
@@ -63,13 +63,19 @@ final class AndroidOptionsInitializer {
       final @NotNull ILogger logger,
       final boolean isFragmentAvailable,
       final boolean isTimberAvailable) {
-    init(options, context, logger, new BuildInfoProvider(), isFragmentAvailable, isTimberAvailable);
+    init(
+        options,
+        context,
+        logger,
+        new BuildInfoProvider(logger),
+        isFragmentAvailable,
+        isTimberAvailable);
   }
 
   /**
    * Init method of the Android Options initializer
    *
-   * @param options the SentryOptions
+   * @param options the SentryAndroidOptions
    * @param context the Application context
    * @param logger the ILogger interface
    * @param buildInfoProvider the BuildInfoProvider interface
@@ -96,7 +102,7 @@ final class AndroidOptionsInitializer {
   /**
    * Init method of the Android Options initializer
    *
-   * @param options the SentryOptions
+   * @param options the SentryAndroidOptions
    * @param context the Application context
    * @param logger the ILogger interface
    * @param buildInfoProvider the BuildInfoProvider interface
@@ -145,11 +151,13 @@ final class AndroidOptionsInitializer {
     options.addEventProcessor(new PerformanceAndroidEventProcessor(options, activityFramesTracker));
 
     options.setTransportGate(new AndroidTransportGate(context, options.getLogger()));
+    options.setTransactionProfiler(
+        new AndroidTransactionProfiler(context, options, buildInfoProvider));
   }
 
   private static void installDefaultIntegrations(
       final @NotNull Context context,
-      final @NotNull SentryOptions options,
+      final @NotNull SentryAndroidOptions options,
       final @NotNull BuildInfoProvider buildInfoProvider,
       final @NotNull LoadClass loadClass,
       final @NotNull ActivityFramesTracker activityFramesTracker,
@@ -191,6 +199,8 @@ final class AndroidOptionsInitializer {
       if (isFragmentAvailable) {
         options.addIntegration(new FragmentLifecycleIntegration((Application) context, true, true));
       }
+      options.addEventProcessor(
+          new ScreenshotEventProcessor((Application) context, options, buildInfoProvider));
     } else {
       options
           .getLogger()
@@ -210,7 +220,7 @@ final class AndroidOptionsInitializer {
   /**
    * Reads and sets default option values that are Android specific like release and inApp
    *
-   * @param options the SentryOptions
+   * @param options the SentryAndroidOptions
    * @param context the Android context methods
    */
   private static void readDefaultOptionValues(
@@ -285,12 +295,33 @@ final class AndroidOptionsInitializer {
    * Sets the cache dirs like sentry, outbox and sessions
    *
    * @param context the Application context
-   * @param options the SentryOptions
+   * @param options the SentryAndroidOptions
    */
+  @SuppressWarnings("FutureReturnValueIgnored")
   private static void initializeCacheDirs(
-      final @NotNull Context context, final @NotNull SentryOptions options) {
+      final @NotNull Context context, final @NotNull SentryAndroidOptions options) {
     final File cacheDir = new File(context.getCacheDir(), "sentry");
+    final File profilingTracesDir = new File(cacheDir, "profiling_traces");
     options.setCacheDirPath(cacheDir.getAbsolutePath());
+    options.setProfilingTracesDirPath(profilingTracesDir.getAbsolutePath());
+
+    if (options.isProfilingEnabled()) {
+      profilingTracesDir.mkdirs();
+      final File[] oldTracesDirContent = profilingTracesDir.listFiles();
+
+      options
+          .getExecutorService()
+          .submit(
+              () -> {
+                if (oldTracesDirContent == null) return;
+                // Method trace files are normally deleted at the end of traces, but if that fails
+                // for some
+                // reason we try to clear any old files here.
+                for (File f : oldTracesDirContent) {
+                  FileUtils.deleteRecursively(f);
+                }
+              });
+    }
   }
 
   private static boolean isNdkAvailable(final @NotNull BuildInfoProvider buildInfoProvider) {