+++ sentry-android-core/src/test/java/io/sentry/android/core/ScreenshotEventProcessorTest.kt
@@ -1,0 +1,194 @@
+package io.sentry.android.core
+
+import android.app.Activity
+import android.app.Application
+import android.view.View
+import android.view.Window
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.never
+import com.nhaarman.mockitokotlin2.verify
+import com.nhaarman.mockitokotlin2.whenever
+import io.sentry.Attachment
+import io.sentry.MainEventProcessor
+import io.sentry.SentryEvent
+import io.sentry.TypeCheckHint.ANDROID_ACTIVITY
+import io.sentry.TypeCheckHint.SENTRY_SCREENSHOT
+import org.junit.runner.RunWith
+import kotlin.test.BeforeTest
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNull
+import kotlin.test.assertSame
+import kotlin.test.assertTrue
+
+@RunWith(AndroidJUnit4::class)
+class ScreenshotEventProcessorTest {
+
+    private class Fixture {
+        val application = mock<Application>()
+        val buildInfo = mock<BuildInfoProvider>()
+        val activity = mock<Activity>()
+        val window = mock<Window>()
+        val view = mock<View>()
+        val rootView = mock<View>()
+        val options = SentryAndroidOptions().apply {
+            dsn = "https://key@sentry.io/proj"
+        }
+        val mainProcessor = MainEventProcessor(options)
+
+        init {
+            whenever(rootView.width).thenReturn(1)
+            whenever(rootView.height).thenReturn(1)
+            whenever(view.rootView).thenReturn(rootView)
+            whenever(window.decorView).thenReturn(view)
+            whenever(activity.window).thenReturn(window)
+        }
+
+        fun getSut(attachScreenshot: Boolean = false): ScreenshotEventProcessor {
+            options.isAttachScreenshot = attachScreenshot
+
+            return ScreenshotEventProcessor(application, options, buildInfo)
+        }
+    }
+
+    private lateinit var fixture: Fixture
+
+    @BeforeTest
+    fun `set up`() {
+        fixture = Fixture()
+    }
+
+    @Test
+    fun `when attach screenshot is enabled, registerActivityLifecycleCallbacks`() {
+        fixture.getSut(true)
+
+        verify(fixture.application).registerActivityLifecycleCallbacks(any())
+    }
+
+    @Test
+    fun `when attach screenshot is disabled, does not registerActivityLifecycleCallbacks`() {
+        fixture.getSut(false)
+
+        verify(fixture.application, never()).registerActivityLifecycleCallbacks(any())
+    }
+
+    @Test
+    fun `when close is called and attach screenshot is enabled, unregisterActivityLifecycleCallbacks`() {
+        val sut = fixture.getSut(true)
+
+        sut.close()
+
+        verify(fixture.application).unregisterActivityLifecycleCallbacks(any())
+    }
+
+    @Test
+    fun `when close is called and  attach screenshot is disabled, does not unregisterActivityLifecycleCallbacks`() {
+        val sut = fixture.getSut(false)
+
+        sut.close()
+
+        verify(fixture.application, never()).unregisterActivityLifecycleCallbacks(any())
+    }
+
+    @Test
+    fun `when process is called and attachScreenshot is disabled, does nothing`() {
+        val sut = fixture.getSut(false)
+        val hints = mutableMapOf<String, Any>()
+
+        sut.onActivityCreated(fixture.activity, null)
+
+        val event = fixture.mainProcessor.process(getEvent(), hints)
+        sut.process(event, hints)
+
+        assertNull(hints[SENTRY_SCREENSHOT])
+    }
+
+    @Test
+    fun `when event is not errored, does nothing`() {
+        val sut = fixture.getSut(true)
+        val hints = mutableMapOf<String, Any>()
+
+        sut.onActivityCreated(fixture.activity, null)
+
+        val event = fixture.mainProcessor.process(SentryEvent(), hints)
+        sut.process(event, hints)
+
+        assertNull(hints[SENTRY_SCREENSHOT])
+    }
+
+    @Test
+    fun `when there is not activity, does nothing`() {
+        val sut = fixture.getSut(true)
+        val hints = mutableMapOf<String, Any>()
+
+        val event = fixture.mainProcessor.process(getEvent(), hints)
+        sut.process(event, hints)
+
+        assertNull(hints[SENTRY_SCREENSHOT])
+    }
+
+    @Test
+    fun `when activity is finishing, does nothing`() {
+        val sut = fixture.getSut(true)
+        val hints = mutableMapOf<String, Any>()
+
+        whenever(fixture.activity.isFinishing).thenReturn(true)
+        sut.onActivityCreated(fixture.activity, null)
+
+        val event = fixture.mainProcessor.process(getEvent(), hints)
+        sut.process(event, hints)
+
+        assertNull(hints[SENTRY_SCREENSHOT])
+    }
+
+    @Test
+    fun `when view is zeroed, does nothing`() {
+        val sut = fixture.getSut(true)
+        val hints = mutableMapOf<String, Any>()
+
+        whenever(fixture.rootView.width).thenReturn(0)
+        whenever(fixture.rootView.height).thenReturn(0)
+        sut.onActivityCreated(fixture.activity, null)
+
+        val event = fixture.mainProcessor.process(getEvent(), hints)
+        sut.process(event, hints)
+
+        assertNull(hints[SENTRY_SCREENSHOT])
+    }
+
+    @Test
+    fun `when process is called and attachScreenshot is enabled, add attachment to hints`() {
+        val sut = fixture.getSut(true)
+        val hints = mutableMapOf<String, Any>()
+
+        sut.onActivityCreated(fixture.activity, null)
+
+        val event = fixture.mainProcessor.process(getEvent(), hints)
+        sut.process(event, hints)
+
+        val screenshot = hints[SENTRY_SCREENSHOT]
+        assertTrue(screenshot is Attachment)
+        assertEquals("screenshot.png", screenshot.filename)
+        assertEquals("image/png", screenshot.contentType)
+
+        assertSame(fixture.activity, hints[ANDROID_ACTIVITY])
+    }
+
+    @Test
+    fun `when activity is destroyed, does nothing`() {
+        val sut = fixture.getSut(true)
+        val hints = mutableMapOf<String, Any>()
+
+        sut.onActivityCreated(fixture.activity, null)
+        sut.onActivityDestroyed(fixture.activity)
+
+        val event = fixture.mainProcessor.process(getEvent(), hints)
+        sut.process(event, hints)
+
+        assertNull(hints[SENTRY_SCREENSHOT])
+    }
+
+    private fun getEvent(): SentryEvent = SentryEvent(Throwable("Throwable"))
+}