+++ sentry-spring/src/test/kotlin/io/sentry/spring/tracing/SentryTransactionAdviceTest.kt
@@ -64,7 +64,9 @@ class SentryTransactionAdviceTest {
                 assertThat(it.contexts.trace!!.operation).isEqualTo("bean")
                 assertThat(it.status).isEqualTo(SpanStatus.OK)
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }
 
@@ -75,7 +77,9 @@ class SentryTransactionAdviceTest {
             check {
                 assertThat(it.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }
 
@@ -87,7 +91,9 @@ class SentryTransactionAdviceTest {
                 assertThat(it.transaction).isEqualTo("SampleService.methodWithoutTransactionNameSet")
                 assertThat(it.contexts.trace!!.operation).isEqualTo("op")
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }
 
@@ -108,7 +114,9 @@ class SentryTransactionAdviceTest {
                 assertThat(it.transaction).isEqualTo("ClassAnnotatedSampleService.hello")
                 assertThat(it.contexts.trace!!.operation).isEqualTo("op")
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }
 
@@ -120,7 +128,9 @@ class SentryTransactionAdviceTest {
                 assertThat(it.transaction).isEqualTo("ClassAnnotatedWithOperationSampleService.hello")
                 assertThat(it.contexts.trace!!.operation).isEqualTo("my-op")
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }