+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -1,6 +1,9 @@
 package io.sentry;
 
+import static io.sentry.TypeCheckHint.SENTRY_SCREENSHOT;
+
 import io.sentry.clientreport.DiscardReason;
+import io.sentry.exception.SentryEnvelopeException;
 import io.sentry.hints.DiskFlushNotification;
 import io.sentry.protocol.Contexts;
 import io.sentry.protocol.SentryId;
@@ -57,7 +60,7 @@ public final class SentryClient implements ISentryClient {
   }
 
   private boolean shouldApplyScopeData(
-      final @NotNull SentryBaseEvent event, final @Nullable Map<String, Object> hint) {
+      final @NotNull SentryBaseEvent event, final @NotNull Map<String, Object> hint) {
     if (HintUtils.shouldApplyScopeData(hint)) {
       return true;
     } else {
@@ -70,10 +73,12 @@ public final class SentryClient implements ISentryClient {
 
   @Override
   public @NotNull SentryId captureEvent(
-      @NotNull SentryEvent event,
-      final @Nullable Scope scope,
-      final @Nullable Map<String, Object> hint) {
+      @NotNull SentryEvent event, final @Nullable Scope scope, @Nullable Map<String, Object> hint) {
     Objects.requireNonNull(event, "SentryEvent is required.");
+
+    if (hint == null) {
+      hint = new HashMap<>();
+    }
 
     options.getLogger().log(SentryLevel.DEBUG, "Capturing event: %s", event.getEventId());
 
@@ -146,12 +151,12 @@ public final class SentryClient implements ISentryClient {
               ? scope.getTransaction().traceState()
               : null;
       final SentryEnvelope envelope =
-          buildEnvelope(event, getAttachmentsFromScope(scope), session, traceState);
+          buildEnvelope(event, getAttachments(scope, hint), session, traceState, null);
 
       if (envelope != null) {
         transport.send(envelope, hint);
       }
-    } catch (IOException e) {
+    } catch (IOException | SentryEnvelopeException e) {
       options.getLogger().log(SentryLevel.WARNING, e, "Capturing event %s failed.", sentryId);
 
       // if there was an error capturing the event, we return an emptyId
@@ -161,20 +166,33 @@ public final class SentryClient implements ISentryClient {
     return sentryId;
   }
 
-  private @Nullable List<Attachment> getAttachmentsFromScope(@Nullable Scope scope) {
+  private @Nullable List<Attachment> getAttachments(
+      final @Nullable Scope scope, final @NotNull Map<String, Object> hint) {
+    List<Attachment> attachments = null;
     if (scope != null) {
-      return scope.getAttachments();
-    } else {
-      return null;
+      attachments = scope.getAttachments();
     }
+
+    final Object screenshotAttachment = hint.get(SENTRY_SCREENSHOT);
+    if (screenshotAttachment instanceof Attachment) {
+
+      if (attachments == null) {
+        attachments = new ArrayList<>();
+      }
+
+      attachments.add((Attachment) screenshotAttachment);
+    }
+
+    return attachments;
   }
 
   private @Nullable SentryEnvelope buildEnvelope(
       final @Nullable SentryBaseEvent event,
       final @Nullable List<Attachment> attachments,
       final @Nullable Session session,
-      final @Nullable TraceState traceState)
-      throws IOException {
+      final @Nullable TraceState traceState,
+      final @Nullable ProfilingTraceData profilingTraceData)
+      throws IOException, SentryEnvelopeException {
     SentryId sentryId = null;
 
     final List<SentryEnvelopeItem> envelopeItems = new ArrayList<>();
@@ -190,6 +208,13 @@ public final class SentryClient implements ISentryClient {
       final SentryEnvelopeItem sessionItem =
           SentryEnvelopeItem.fromSession(options.getSerializer(), session);
       envelopeItems.add(sessionItem);
+    }
+
+    if (profilingTraceData != null) {
+      final SentryEnvelopeItem profilingTraceItem =
+          SentryEnvelopeItem.fromProfilingTrace(
+              profilingTraceData, options.getMaxTraceFileSize(), options.getSerializer());
+      envelopeItems.add(profilingTraceItem);
     }
 
     if (attachments != null) {
@@ -212,7 +237,7 @@ public final class SentryClient implements ISentryClient {
   @Nullable
   private SentryEvent processEvent(
       @NotNull SentryEvent event,
-      final @Nullable Map<String, Object> hint,
+      final @NotNull Map<String, Object> hint,
       final @NotNull List<EventProcessor> eventProcessors) {
     for (final EventProcessor processor : eventProcessors) {
       try {
@@ -246,7 +271,7 @@ public final class SentryClient implements ISentryClient {
   @Nullable
   private SentryTransaction processTransaction(
       @NotNull SentryTransaction transaction,
-      final @Nullable Map<String, Object> hint,
+      final @NotNull Map<String, Object> hint,
       final @NotNull List<EventProcessor> eventProcessors) {
     for (final EventProcessor processor : eventProcessors) {
       try {
@@ -424,8 +449,13 @@ public final class SentryClient implements ISentryClient {
       @NotNull SentryTransaction transaction,
       @Nullable TraceState traceState,
       final @Nullable Scope scope,
-      final @Nullable Map<String, Object> hint) {
+      @Nullable Map<String, Object> hint,
+      final @Nullable ProfilingTraceData profilingTraceData) {
     Objects.requireNonNull(transaction, "Transaction is required.");
+
+    if (hint == null) {
+      hint = new HashMap<>();
+    }
 
     options
         .getLogger()
@@ -460,13 +490,18 @@ public final class SentryClient implements ISentryClient {
     try {
       final SentryEnvelope envelope =
           buildEnvelope(
-              transaction, filterForTransaction(getAttachmentsFromScope(scope)), null, traceState);
+              transaction,
+              filterForTransaction(getAttachments(scope, hint)),
+              null,
+              traceState,
+              profilingTraceData);
+
       if (envelope != null) {
         transport.send(envelope, hint);
       } else {
         sentryId = SentryId.EMPTY_ID;
       }
-    } catch (IOException e) {
+    } catch (IOException | SentryEnvelopeException e) {
       options.getLogger().log(SentryLevel.WARNING, e, "Capturing transaction %s failed.", sentryId);
       // if there was an error capturing the event, we return an emptyId
       sentryId = SentryId.EMPTY_ID;
@@ -493,7 +528,7 @@ public final class SentryClient implements ISentryClient {
   private @Nullable SentryEvent applyScope(
       @NotNull SentryEvent event,
       final @Nullable Scope scope,
-      final @Nullable Map<String, Object> hint) {
+      final @NotNull Map<String, Object> hint) {
     if (scope != null) {
       applyScope(event, scope);
 
@@ -571,7 +606,7 @@ public final class SentryClient implements ISentryClient {
   }
 
   private @Nullable SentryEvent executeBeforeSend(
-      @NotNull SentryEvent event, final @Nullable Map<String, Object> hint) {
+      @NotNull SentryEvent event, final @NotNull Map<String, Object> hint) {
     final SentryOptions.BeforeSendCallback beforeSend = options.getBeforeSend();
     if (beforeSend != null) {
       try {