+++ sentry/src/main/java/io/sentry/SentryTracer.java
@@ -211,6 +211,14 @@ public final class SentryTracer implements ITransaction {
   public void finish(@Nullable SpanStatus status) {
     this.finishStatus = FinishStatus.finishing(status);
     if (!root.isFinished() && (!waitForChildren || hasAllChildrenFinished())) {
+      ProfilingTraceData profilingTraceData = null;
+      Boolean isSampled = isSampled();
+      if (isSampled == null) {
+        isSampled = false;
+      }
+      if (hub.getOptions().isProfilingEnabled() && isSampled) {
+        profilingTraceData = hub.getOptions().getTransactionProfiler().onTransactionFinish(this);
+      }
       root.finish(finishStatus.spanStatus);
 
       // finish unfinished children
@@ -244,7 +252,7 @@ public final class SentryTracer implements ITransaction {
       if (transactionFinishedCallback != null) {
         transactionFinishedCallback.execute(this);
       }
-      hub.captureTransaction(transaction, this.traceState());
+      hub.captureTransaction(transaction, this.traceState(), null, profilingTraceData);
     }
   }