+++ sentry/src/test/java/io/sentry/NoOpTransactionProfilerTest.kt
@@ -1,0 +1,16 @@
+package io.sentry
+
+import kotlin.test.Test
+import kotlin.test.assertNull
+
+class NoOpTransactionProfilerTest {
+    private var profiler = NoOpTransactionProfiler.getInstance()
+
+    @Test
+    fun `onTransactionStart does not throw`() =
+        profiler.onTransactionStart(NoOpTransaction.getInstance())
+
+    @Test
+    fun `onTransactionFinish returns null`() =
+        assertNull(profiler.onTransactionFinish(NoOpTransaction.getInstance()))
+}