+++ sentry/src/test/java/io/sentry/protocol/ContextsTest.kt
@@ -1,7 +1,7 @@
 package io.sentry.protocol
 
 import io.sentry.SpanContext
-import org.junit.Test
+import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertNotSame